+++ pkg/store/bucket_test.go
@@ -12,9 +12,149 @@ import (
 	"github.com/improbable-eng/thanos/pkg/compact/downsample"
 	"github.com/improbable-eng/thanos/pkg/store/storepb"
 	"github.com/improbable-eng/thanos/pkg/testutil"
+	"github.com/leanovate/gopter"
+	"github.com/leanovate/gopter/gen"
+	"github.com/leanovate/gopter/prop"
 	"github.com/oklog/ulid"
 	"github.com/prometheus/tsdb/labels"
 )
+
+func TestBucketBlock_Property(t *testing.T) {
+	parameters := gopter.DefaultTestParameters()
+	parameters.Rng.Seed(2000)
+	parameters.MinSuccessfulTests = 20000
+	properties := gopter.NewProperties(parameters)
+
+	set := newBucketBlockSet(labels.Labels{})
+
+	type resBlock struct {
+		mint, maxt int64
+		window     int64
+	}
+	// This input resembles a typical production-level block layout
+	// in remote object storage.
+	input := []resBlock{
+		{window: downsample.ResLevel0, mint: 0, maxt: 100},
+		{window: downsample.ResLevel0, mint: 100, maxt: 200},
+		// Compaction level 2 begins but not downsampling (8 hour block length)
+		{window: downsample.ResLevel0, mint: 200, maxt: 600},
+		{window: downsample.ResLevel0, mint: 600, maxt: 1000},
+		// Compaction level 3 begins, Some of it is downsampled but still retained (48 hour block length)
+		{window: downsample.ResLevel0, mint: 1000, maxt: 1750},
+		{window: downsample.ResLevel1, mint: 1000, maxt: 1750},
+		// Compaction level 4 begins, different downsampling levels cover the same (336 hour block length)
+		{window: downsample.ResLevel0, mint: 1750, maxt: 7000},
+		{window: downsample.ResLevel1, mint: 1750, maxt: 7000},
+		{window: downsample.ResLevel2, mint: 1750, maxt: 7000},
+		// Compaction level 4 already happened, raw samples have been deleted
+		{window: downsample.ResLevel0, mint: 7000, maxt: 14000},
+		{window: downsample.ResLevel1, mint: 7000, maxt: 14000},
+		// Compaction level 4 already happened, raw and downsample res level 1 samples have been deleted
+		{window: downsample.ResLevel2, mint: 14000, maxt: 21000},
+	}
+
+	for _, in := range input {
+		var m metadata.Meta
+		m.Thanos.Downsample.Resolution = in.window
+		m.MinTime = in.mint
+		m.MaxTime = in.maxt
+
+		testutil.Ok(t, set.add(&bucketBlock{meta: &m}))
+	}
+
+	properties.Property("getFor always gets at least some data in range", prop.ForAllNoShrink(
+		func(low, high, maxResolution int64) bool {
+			// Bogus case.
+			if low >= high {
+				return true
+			}
+
+			res := set.getFor(low, high, maxResolution)
+
+			// The data that we get must all encompass our requested range
+			if len(res) == 1 && (res[0].meta.Thanos.Downsample.Resolution > maxResolution ||
+				res[0].meta.MinTime > low) {
+				return false
+			} else if len(res) > 1 {
+				mint := int64(21001)
+				maxt := int64(0)
+				for i := 0; i < len(res)-1; i++ {
+					if res[i].meta.Thanos.Downsample.Resolution > maxResolution {
+						return false
+					}
+					if res[i+1].meta.MinTime != res[i].meta.MaxTime {
+						return false
+					}
+					if res[i].meta.MinTime < mint {
+						mint = res[i].meta.MinTime
+					}
+					if res[i].meta.MaxTime > maxt {
+						maxt = res[i].meta.MaxTime
+					}
+				}
+				if res[len(res)-1].meta.MinTime < mint {
+					mint = res[len(res)-1].meta.MinTime
+				}
+				if res[len(res)-1].meta.MaxTime > maxt {
+					maxt = res[len(res)-1].meta.MaxTime
+				}
+				if low < mint {
+					return false
+				}
+
+			}
+			return true
+		}, gen.Int64Range(0, 21000), gen.Int64Range(0, 21000), gen.Int64Range(0, 60*60*1000)),
+	)
+
+	properties.Property("getFor always gets all data in range", prop.ForAllNoShrink(
+		func(low, high int64) bool {
+			// Bogus case.
+			if low >= high {
+				return true
+			}
+
+			maxResolution := downsample.ResLevel2
+			res := set.getFor(low, high, maxResolution)
+
+			// The data that we get must all encompass our requested range
+			if len(res) == 1 && (res[0].meta.Thanos.Downsample.Resolution > maxResolution ||
+				res[0].meta.MinTime > low || res[0].meta.MaxTime < high) {
+				return false
+			} else if len(res) > 1 {
+				mint := int64(21001)
+				maxt := int64(0)
+				for i := 0; i < len(res)-1; i++ {
+					if res[i+1].meta.MinTime != res[i].meta.MaxTime {
+						return false
+					}
+					if res[i].meta.MinTime < mint {
+						mint = res[i].meta.MinTime
+					}
+					if res[i].meta.MaxTime > maxt {
+						maxt = res[i].meta.MaxTime
+					}
+				}
+				if res[len(res)-1].meta.MinTime < mint {
+					mint = res[len(res)-1].meta.MinTime
+				}
+				if res[len(res)-1].meta.MaxTime > maxt {
+					maxt = res[len(res)-1].meta.MaxTime
+				}
+				if low < mint {
+					return false
+				}
+				if high > maxt {
+					return false
+				}
+
+			}
+			return true
+		}, gen.Int64Range(0, 21000), gen.Int64Range(0, 21000)),
+	)
+
+	properties.TestingRun(t)
+}
 
 func TestBucketBlockSet_addGet(t *testing.T) {
 	defer leaktest.CheckTimeout(t, 10*time.Second)()
@@ -53,13 +193,13 @@ func TestBucketBlockSet_addGet(t *testing.T) {
 
 	cases := []struct {
 		mint, maxt    int64
-		minResolution int64
+		maxResolution int64
 		res           []resBlock
 	}{
 		{
 			mint:          -100,
 			maxt:          1000,
-			minResolution: 0,
+			maxResolution: 0,
 			res: []resBlock{
 				{window: downsample.ResLevel0, mint: 0, maxt: 100},
 				{window: downsample.ResLevel0, mint: 100, maxt: 200},
@@ -70,7 +210,7 @@ func TestBucketBlockSet_addGet(t *testing.T) {
 		}, {
 			mint:          100,
 			maxt:          400,
-			minResolution: downsample.ResLevel1 - 1,
+			maxResolution: downsample.ResLevel1 - 1,
 			res: []resBlock{
 				{window: downsample.ResLevel0, mint: 100, maxt: 200},
 				{window: downsample.ResLevel0, mint: 200, maxt: 300},
@@ -79,7 +219,7 @@ func TestBucketBlockSet_addGet(t *testing.T) {
 		}, {
 			mint:          100,
 			maxt:          500,
-			minResolution: downsample.ResLevel1,
+			maxResolution: downsample.ResLevel1,
 			res: []resBlock{
 				{window: downsample.ResLevel1, mint: 100, maxt: 200},
 				{window: downsample.ResLevel1, mint: 200, maxt: 300},
@@ -89,7 +229,7 @@ func TestBucketBlockSet_addGet(t *testing.T) {
 		}, {
 			mint:          0,
 			maxt:          500,
-			minResolution: downsample.ResLevel2,
+			maxResolution: downsample.ResLevel2,
 			res: []resBlock{
 				{window: downsample.ResLevel1, mint: 0, maxt: 100},
 				{window: downsample.ResLevel2, mint: 100, maxt: 200},
@@ -110,7 +250,7 @@ func TestBucketBlockSet_addGet(t *testing.T) {
 			m.MaxTime = b.maxt
 			exp = append(exp, &bucketBlock{meta: &m})
 		}
-		res := set.getFor(c.mint, c.maxt, c.minResolution)
+		res := set.getFor(c.mint, c.maxt, c.maxResolution)
 		testutil.Equals(t, exp, res)
 	}
 }